+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 26.5.17
 * Add base salary row
+* Fix specialty bonus so it only applies to the skill-derived wage, not the fixed base salary (ref: https://www.hattrick.org/Forum/Read.aspx?t=17629498&v=4&a=1&n=237)
 
 ## 26.3.12
 * Better support of multiple teams